生产领料做完

+++ src/views/prdissue/PrdIssueList.vue
@@ -5,6 +5,15 @@
      <!--插槽:table标题-->
       <template #tableTitle>
           <a-button type="primary" @click="handleAdd" preIcon="ant-design:plus-outlined"> 新增</a-button>
+        <!-- 审核/反审核按钮 -->
+        <a-button type="primary" preIcon="ant-design:check-circle-twotone"
+                  @click="onAudit('audit', selectedRowKeys)"
+                  :disabled="selectedRowKeys.length === 0">审核
+        </a-button>
+        <a-button type="primary" preIcon="ant-design:close-circle-twotone"
+                  @click="onAudit('reverse', selectedRowKeys)"
+                  :disabled="selectedRowKeys.length === 0">反审核
+        </a-button>
           <a-button  type="primary" preIcon="ant-design:export-outlined" @click="onExportXls"> 导出</a-button>
           <j-upload-button  type="primary" preIcon="ant-design:import-outlined" @click="onImportXls">导入</j-upload-button>
           <a-dropdown v-if="selectedRowKeys.length > 0">
@@ -43,16 +52,24 @@
   import {useModal} from '/@/components/Modal';
   import PrdIssueModal from './components/PrdIssueModal.vue'
   import {columns, searchFormSchema, superQuerySchema} from './PrdIssue.data';
-  import {list, deleteOne, batchDelete, getImportUrl,getExportUrl} from './PrdIssue.api';
+  import {
+    list,
+    deleteOne,
+    batchDelete,
+    getImportUrl,
+    getExportUrl,
+    getAuditUrl
+  } from './PrdIssue.api';
   import {downloadFile} from '/@/utils/common/renderUtils';
   import { useUserStore } from '/@/store/modules/user';
+  import {useMessage} from "@/hooks/web/useMessage";
   const queryParam = reactive<any>({});
   const checkedKeys = ref<Array<string | number>>([]);
   const userStore = useUserStore();
   //注册model
   const [registerModal, {openModal}] = useModal();
    //注册table数据
-  const { prefixCls,tableContext,onExportXls,onImportXls } = useListPage({
+  const { prefixCls,tableContext,onExportXls,onImportXls ,onAudit} = useListPage({
       tableProps:{
            title: '生产领料',
            api: list,
@@ -85,13 +102,17 @@
             url: getImportUrl,
             success: handleSuccess
         },
+    auditConfig: {
+      url: getAuditUrl,
+      success: handleSuccess
+    }
     })
 
   const [registerTable, {reload},{ rowSelection, selectedRowKeys }] = tableContext
 
   // 高级查询配置
   const superQueryConfig = reactive(superQuerySchema);
-
+  const { createMessage } = useMessage()
   /**
    * 高级查询事件
    */
@@ -115,6 +136,10 @@
     * 编辑事件
     */
   function handleEdit(record: Recordable) {
+     if (record?.audit === 1) { // 如果 audit 是 ref 或对象形式
+       createMessage.warning('已审核单据不能操作，请先反审核！');
+       return;
+     }
      openModal(true, {
        record,
        isUpdate: true,
@@ -135,6 +160,10 @@
     * 删除事件
     */
   async function handleDelete(record) {
+     if (record?.audit === 1) { // 如果 audit 是 ref 或对象形式
+       createMessage.warning('已审核单据不能操作，请先反审核！');
+       return;
+     }
      await deleteOne({id: record.id}, handleSuccess);
    }
    /**
@@ -180,7 +209,7 @@
       }
     ]
   }
-
+  window.handleDetail=handleDetail
 </script>
 
 <style scoped>
